Update and rename firebase.js to firebase.ts
